requirements.txt
flask
flask-swagger-ui
pyyaml
pandas
flask-restx
gevent
werkzeug